silence block corner cases

+++ js/blocks.js
@@ -1320,9 +1320,15 @@ class Blocks {
             }
 
             let thisBlockobj = this.blockList[thisBlock];
-            if (thisBlockobj.name === "vspace") {
+
+            // Do not remove the silence block if only a vspace block is added before the silence block.
+            if (
+                thisBlockobj.name === "vspace" &&
+                this.blockList[thisBlockobj.connections[1]].name === "vspace" &&
+                this.blockList[thisBlockobj.connections[0]].name === "newnote"
+              ) {
                 return;
-            }
+              }
 
             thisBlockobj = this.blockList[thisBlock];
             if (thisBlockobj.name === "rest2") {
@@ -1364,6 +1370,12 @@ class Blocks {
          */
         this.deletePreviousDefault = (thisBlock) => {
             let thisBlockobj = this.blockList[thisBlock];
+
+            // Do not remove the silence block if only a vspace block is inserted after the silence block.
+            if (this.blockList[thisBlockobj.connections[0]]?.name === "rest2" && this.blockList[thisBlock]?.name === "vspace") {
+                return thisBlockobj.connections[0];
+            }
+
             if (this._blockInStack(thisBlock, ["rest2"])) {
                 this._deletePitchBlocks(thisBlock);
                 return this.blockList[thisBlock].connections[0];
